Some refactoring and start of negotiation implementation.
