+++ src/main/java/org/reaktivity/nukleus/socks/internal/stream/server/ServerStreamFactory.java
@@ -1,0 +1,124 @@
+/**
+ * Copyright 2016-2017 The Reaktivity Project
+ *
+ * The Reaktivity Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package org.reaktivity.nukleus.socks.internal.stream.server;
+
+import java.util.function.LongSupplier;
+
+import org.agrona.DirectBuffer;
+import org.agrona.MutableDirectBuffer;
+import org.agrona.collections.Long2ObjectHashMap;
+import org.reaktivity.nukleus.Configuration;
+import org.reaktivity.nukleus.buffer.BufferPool;
+import org.reaktivity.nukleus.function.MessageConsumer;
+import org.reaktivity.nukleus.function.MessagePredicate;
+import org.reaktivity.nukleus.route.RouteManager;
+import org.reaktivity.nukleus.socks.internal.stream.Correlation;
+import org.reaktivity.nukleus.socks.internal.types.control.RouteFW;
+import org.reaktivity.nukleus.socks.internal.types.stream.BeginFW;
+import org.reaktivity.nukleus.stream.StreamFactory;
+
+public final class ServerStreamFactory implements StreamFactory
+{
+
+    private final StreamContext streamContext;
+
+    public ServerStreamFactory(
+        Configuration config,
+        RouteManager router,
+        MutableDirectBuffer writeBuffer,
+        BufferPool bufferPool,
+        LongSupplier supplyStreamId,
+        LongSupplier supplyCorrelationId,
+        Long2ObjectHashMap<Correlation> correlations)
+    {
+        this.streamContext =
+            new StreamContext(
+                config,
+                router,
+                writeBuffer,
+                bufferPool,
+                supplyStreamId,
+                supplyCorrelationId,
+                correlations,
+                this::wrapRoute);
+    }
+
+    @Override
+    public MessageConsumer newStream(
+        int msgTypeId,
+        DirectBuffer buffer,
+        int index,
+        int length,
+        MessageConsumer throttle)
+    {
+        final BeginFW begin = streamContext.beginRO.wrap(buffer, index, index + length);
+        final long sourceRef = begin.sourceRef();
+
+        if (sourceRef == 0L)
+        {
+            System.out.println("newConnectReplyStream");
+            return newConnectReplyStream(begin, throttle);
+        }
+        else
+        {
+            System.out.println("newAcceptStream");
+            return newAcceptStream(begin, throttle);
+        }
+    }
+
+    private MessageConsumer newAcceptStream(
+        final BeginFW begin,
+        final MessageConsumer acceptThrottle)
+    {
+        final long acceptRef = begin.sourceRef();
+        final String acceptName = begin.source()
+            .asString();
+        final MessagePredicate filter = (t, b, o, l) ->
+        {
+            final RouteFW route = streamContext.routeRO.wrap(b, o, l);
+            return acceptRef == route.sourceRef() && acceptName.equals(route.source()
+                .asString());
+        };
+
+        final RouteFW route = streamContext.router.resolve(filter, this::wrapRoute);
+        if (route != null)
+        {
+            final long acceptId = begin.streamId();
+            System.out.println("Creating new Server accept Stream");
+            return new AcceptStreamHandler(streamContext, acceptThrottle, acceptId, acceptRef)::handleStream;
+        }
+        return null;
+    }
+
+    private MessageConsumer newConnectReplyStream(
+        final BeginFW begin,
+        final MessageConsumer connectReplyThrottle)
+    {
+        final long connectReplyId = begin.streamId();
+        return new ConnectReplyStreamHandler(streamContext, connectReplyThrottle, connectReplyId)::handleStream;
+    }
+
+    private RouteFW wrapRoute(
+        int msgTypeId,
+        DirectBuffer buffer,
+        int index,
+        int length)
+    {
+        System.out.println("WrapRoute");
+        return streamContext.routeRO.wrap(buffer, index, index + length);
+    }
+
+}
